+++ R/tar_path.R
@@ -3,14 +3,27 @@
 #' @family utilities
 #' @description Identify the file path where a target will be stored
 #'   after the target finishes running in the pipeline.
-#' @details `tar_path(name = your_target)` just returns
+#' @return Character, file path of the return value of the target.
+#'   If not called from inside a running target,
+#'   `tar_path(name = your_target)` just returns
 #'   `_targets/objects/your_target`, the file path where `your_target`
 #'   will be saved unless `format` is equal to `"file"` or any of the
-#'   supported cloud-based storage formats. If you call `tar_path()`
+#'   supported cloud-based storage formats.
+#'   
+#'   For non-cloud storage formats, if you call `tar_path()`
 #'   with no arguments while target `x` is running, the `name`
-#'   argument defaults to the name of the target,
+#'   argument defaults to the name of the running target,
 #'   so `tar_path()` returns `_targets/objects/x`.
-#' @return Character, file path to a hypothetical target.
+#'
+#'   For cloud-backed formats, `tar_path()` returns the
+#'   path to the staging file in `_targets/scratch/`.
+#'   That way, even if you select a cloud format
+#'   (e.g. `tar_target(..., format = "aws_parquet", storage = "none")`)
+#'   then you can still manually write to `tar_path(create_dir = TRUE)`
+#'   and the `targets` package will automatically hash it and
+#'   upload it to the AWS S3 bucket. This does not apply to formats
+#'   `"file"` or `"aws_file"`, where you would never need `storage = "none"`
+#'   anyway.
 #' @inheritParams tar_validate
 #' @param name Symbol, name of a target.
 #'   If `NULL`, `tar_path()` returns the path of the target currently running
@@ -19,6 +32,11 @@
 #'   is called on its own outside a `targets` pipeline.
 #'   Having a default lets users run things without [tar_make()],
 #'   which helps peel back layers of code and troubleshoot bugs.
+#' @param create_dir Logical of length 1,
+#'   whether to create `dirname(tar_path())` in `tar_path()` itself.
+#'   This is useful if you are writing to `tar_path()` from inside a
+#'   `storage = "none"` target and need the parent directory of the file
+#'   to exist.
 #' @examples
 #' tar_path()
 #' tar_path(your_target)
@@ -32,16 +50,23 @@
 tar_path <- function(
   name = NULL,
   default = NA_character_,
+  create_dir = FALSE,
   store = targets::tar_config_get("store")
 ) {
   name <- tar_deparse_language(substitute(name))
   tar_assert_chr(name %|||% character(0))
   tar_assert_chr(default)
-  if_any(
+  tar_assert_lgl(create_dir)
+  tar_assert_scalar(create_dir)
+  out <- if_any(
     is.null(name),
     tar_path_running(default, path_store = store),
     path_objects(path_store = store, name = name)
   )
+  if (create_dir) {
+    dir_create(dirname(out))
+  }
+  out
 }
 
 tar_path_running <- function(default, path_store) {